+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [1.0.1] - 23/03/2019
+
+* Use snapToEnd to prevent intro animation on widget creation
+
+## [1.0.0] - 23/03/2019
+
+* Add animations name customization
+
 ## [0.0.1+2] - 27/02/2019
 
 * Add width and height field.